simple initialize
