+++ app/src/main/java/chilladvanced/ExoPlayerWithStat.kt
@@ -1,15 +1,22 @@
 package chilladvanced
 
+import android.net.Uri
 import com.google.android.exoplayer2.ExoPlayer
+import com.google.android.exoplayer2.MediaItem
+import com.google.android.exoplayer2.source.MediaSource
+
 
 class ExoPlayerWithStat(val exoPlayer: ExoPlayer) : ExoPlayer by exoPlayer {
     override fun prepare() {
-        // TODO
+
+
+
+        exoPlayer.prepare()
+        val mediaSource: MediaItem? = exoPlayer.currentMediaItem
         Logger.register(
-            exoPlayer.currentMediaItem?.requestMetadata?.mediaUri?.host ?: "",
+            mediaSource?.requestMetadata?.mediaUri?.host ?: " ",
             NETWORK_METHODS.EXO_PLAYER
         )
-        exoPlayer.prepare()
     }
